snap-recovery: rename from recovery-tool and update tests
